requirements.txt
apache-beam==2.48.0
apache-flink==1.18.0
apache-flink-libraries==1.18.0
attrs==25.3.0
avro==1.12.0
avro-python3==1.9.1
cassandra-driver==3.28.0
certifi==2025.1.31
charset-normalizer==3.4.1
cloudpickle==2.2.1
confluent-kafka==2.3.0
crcmod==1.7
dill==0.3.1.1
dnspython==2.7.0
docopt==0.6.2
fastavro==1.10.0
fasteners==0.19
find_libpython==0.4.0
geomet==0.2.1.post1
grpcio==1.65.5
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
jsonpickle==3.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
numpy==1.24.4
objsize==0.6.1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pemja==0.3.0
proto-plus==1.26.1
protobuf==4.23.4
py4j==0.10.9.7
pyarrow==11.0.0
pyarrow-hotfix==0.6
pydot==1.4.2
pymongo==4.11.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
six==1.17.0
sortedcontainers==2.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
zstandard==0.23.0
